libdpkg: Fix integer overflow in unit test
Fixes: commit 59f63342b9121b9d941d3dbd09487c953a113f6e
(cherry picked from commit 0fe059239ea00e6c82830f15a9417889d2789839)

+++ lib/dpkg/t/t-deb-version.c
@@ -74,7 +74,7 @@ test_deb_version_parse(void)
 	test_pass(deb_version_parse(&v, vs) == NULL);
 	free(vs);
 
-	if (asprintf(&vs, "%ld.0", (long int)(1L + INT_MAX)) < 0)
+	if (asprintf(&vs, "%u.0", 1U + (unsigned int)INT_MAX) < 0)
 		test_bail("cannot allocate memory for asprintf()");
 	test_fail(deb_version_parse(&v, vs) == NULL);
 	free(vs);
